requirements.txt
plyer>=2.1
click>=8.1.0
pyyaml>=6.0